refactor: CHANGE HEART, SCRAP DIR (#67)
