--- src/main/java/paratrip/paratrip/board/hearts/service/dto/request/BoardHeartRequestDto.java
@@ -1,18 +1,0 @@
-package paratrip.paratrip.board.hearts.service.dto.request;
-
-public class BoardHeartRequestDto {
-	public record IncreaseBoardHeartsRequestDto(
-		Long memberSeq,
-		Long boardSeq
-	) {
-
-	}
-
-	public record DecreaseBoardHeartsRequestDto(
-		Long memberSeq,
-		Long boardSeq,
-		Long boardHeartSeq
-	) {
-
-	}
-}
