--- src/main/java/paratrip/paratrip/board/hearts/service/dto/response/BoardHeartResponseDto.java
@@ -1,9 +1,0 @@
-package paratrip.paratrip.board.hearts.service.dto.response;
-
-public class BoardHeartResponseDto {
-	public record AddBoardHeartResponseDto(
-		Long boardHeartSeq
-	) {
-
-	}
-}
