+++ src/main/java/paratrip/paratrip/board/main/service/BoardService.java
@@ -11,7 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import lombok.RequiredArgsConstructor;
-import paratrip.paratrip.board.hearts.repoisitory.BoardHeartRepository;
+import paratrip.paratrip.hearts.repoisitory.BoardHeartRepository;
 import paratrip.paratrip.board.main.domain.BoardDomain;
 import paratrip.paratrip.board.main.entity.BoardEntity;
 import paratrip.paratrip.board.main.entity.BoardImageEntity;
@@ -21,7 +21,7 @@ import paratrip.paratrip.board.main.repository.BoardImageRepository;
 import paratrip.paratrip.board.main.repository.BoardRepository;
 import paratrip.paratrip.board.main.service.dto.request.BoardRequestDto;
 import paratrip.paratrip.board.main.service.dto.response.BoardResponseDto;
-import paratrip.paratrip.board.scrap.repository.BoardScrapRepository;
+import paratrip.paratrip.scrap.repository.BoardScrapRepository;
 import paratrip.paratrip.board.search.mapper.BoardDocumentsMapper;
 import paratrip.paratrip.board.search.repository.BoardDocumentsRepository;
 import paratrip.paratrip.comment.repository.CommentRepository;
