--- src/main/java/paratrip/paratrip/board/scrap/controller/vo/response/BoardScrapResponseVo.java
@@ -1,9 +1,0 @@
-package paratrip.paratrip.board.scrap.controller.vo.response;
-
-public class BoardScrapResponseVo {
-	public record AddBoardScrapResponse(
-		Long boardScrapResponse
-	) {
-
-	}
-}
