--- src/main/java/paratrip/paratrip/board/scrap/entity/BoardScrapEntity.java
@@ -1,43 +1,0 @@
-package paratrip.paratrip.board.scrap.entity;
-
-import org.hibernate.annotations.DynamicInsert;
-import org.hibernate.annotations.DynamicUpdate;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import paratrip.paratrip.board.main.entity.BoardEntity;
-import paratrip.paratrip.member.entity.MemberEntity;
-
-@Entity
-@Table(name = "BOARD_SCRAP")
-@AllArgsConstructor
-@NoArgsConstructor
-@DynamicInsert
-@DynamicUpdate
-@Getter
-@Builder(toBuilder = true)
-public class BoardScrapEntity {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "board_scrap_seq")
-	private Long boardScrapSeq;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "member_seq")
-	private MemberEntity memberEntity;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "board_seq")
-	private BoardEntity boardEntity;
-}
