+++ src/main/java/paratrip/paratrip/board/search/service/BoardDocumentsService.java
@@ -11,7 +11,7 @@ import lombok.RequiredArgsConstructor;
 import paratrip.paratrip.board.main.entity.BoardEntity;
 import paratrip.paratrip.board.main.repository.BoardImageRepository;
 import paratrip.paratrip.board.main.repository.BoardRepository;
-import paratrip.paratrip.board.scrap.repository.BoardScrapRepository;
+import paratrip.paratrip.scrap.repository.BoardScrapRepository;
 import paratrip.paratrip.board.search.repository.BoardDocumentsRepository;
 import paratrip.paratrip.board.search.service.dto.response.BoardDocumentsResponseDto;
 import paratrip.paratrip.comment.repository.CommentRepository;
