+++ src/main/java/paratrip/paratrip/hearts/service/dto/request/BoardHeartRequestDto.java
@@ -1,0 +1,18 @@
+package paratrip.paratrip.hearts.service.dto.request;
+
+public class BoardHeartRequestDto {
+	public record IncreaseBoardHeartsRequestDto(
+		Long memberSeq,
+		Long boardSeq
+	) {
+
+	}
+
+	public record DecreaseBoardHeartsRequestDto(
+		Long memberSeq,
+		Long boardSeq,
+		Long boardHeartSeq
+	) {
+
+	}
+}
