+++ src/main/java/paratrip/paratrip/hearts/service/dto/response/BoardHeartResponseDto.java
@@ -1,0 +1,9 @@
+package paratrip.paratrip.hearts.service.dto.response;
+
+public class BoardHeartResponseDto {
+	public record AddBoardHeartResponseDto(
+		Long boardHeartSeq
+	) {
+
+	}
+}
