+++ src/main/java/paratrip/paratrip/scrap/controller/vo/response/BoardScrapResponseVo.java
@@ -1,0 +1,9 @@
+package paratrip.paratrip.scrap.controller.vo.response;
+
+public class BoardScrapResponseVo {
+	public record AddBoardScrapResponse(
+		Long boardScrapResponse
+	) {
+
+	}
+}
