+++ src/main/java/paratrip/paratrip/scrap/service/dto/request/BoardScarpRequestDto.java
@@ -1,0 +1,17 @@
+package paratrip.paratrip.scrap.service.dto.request;
+
+public class BoardScarpRequestDto {
+	public record AddBoardScrapRequestDto(
+		Long memberSeq,
+		Long boardSeq
+	) {
+
+	}
+
+	public record DeleteBoardScrapRequestDto(
+		Long memberSeq,
+		Long boardScrapSeq
+	) {
+
+	}
+}
